+++ tldr_app/tldr-dekstop/tldr_cpp/src/lib_tldr/db/postgres_database.h
@@ -14,7 +14,7 @@ namespace tldr {
         bool initialize() override;
         int64_t saveEmbeddings(const std::vector<std::string_view> &chunks,
                                const json &embeddings_response,
-                               const std::vector<size_t> &embedding_hashes = {}) override;
+                               const std::vector<uint64_t> &embedding_hashes = {}) override;
         bool getEmbeddings(int64_t id, std::vector<std::string> &chunks, json &embeddings) override;
 
         // Perform vector similarity search
